+++ lib/sw_apps/zynqmp_pmufw/src/pm_sram.c
@@ -1,26 +1,8 @@
 /*
- * Copyright (C) 2014 - 2019 Xilinx, Inc.  All rights reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- *
+* Copyright (c) 2014 - 2020 Xilinx, Inc.  All rights reserved.
+* SPDX-License-Identifier: MIT
  */
+
 #include "xpfw_config.h"
 #ifdef ENABLE_PM
 
@@ -59,7 +41,7 @@
 #define PM_SRAM_OFF_TO_ON_LATENCY	3100U
 
 /* Sram states */
-static const u32 pmSramStates[PM_SRAM_STATE_MAX] = {
+static const u8 pmSramStates[PM_SRAM_STATE_MAX] = {
 	[PM_SRAM_STATE_OFF] = 0U,
 	[PM_SRAM_STATE_RET] = PM_CAP_CONTEXT | PM_CAP_POWER,
 	[PM_SRAM_STATE_ON] = PM_CAP_ACCESS | PM_CAP_CONTEXT | PM_CAP_POWER,
@@ -257,7 +239,7 @@ static const PmSlaveFsm pmSlaveTcmFsm = {
 	.enterState = PmTcmFsmHandler,
 };
 
-static u32 PmSramPowers[] = {
+static u8 PmSramPowers[] = {
 	DEFAULT_SRAM_POWER_OFF,
 	DEFAULT_SRAM_POWER_RETENTION,
 	DEFAULT_SRAM_POWER_ON,
