+++ lib/sw_apps/zynqmp_pmufw/src/pm_usb.c
@@ -1,26 +1,8 @@
 /*
- * Copyright (C) 2014 - 2019 Xilinx, Inc.  All rights reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- *
+* Copyright (c) 2014 - 2020 Xilinx, Inc.  All rights reserved.
+* SPDX-License-Identifier: MIT
  */
+
 #include "xpfw_config.h"
 #ifdef ENABLE_PM
 
@@ -51,7 +33,7 @@
 #define PM_USB_OFF_TO_ON_LATENCY	152U
 
 /* USB states */
-static const u32 pmUsbStates[] = {
+static const u8 pmUsbStates[] = {
 	[PM_USB_STATE_UNUSED] = 0U,
 	[PM_USB_STATE_OFF] = PM_CAP_WAKEUP | PM_CAP_POWER,
 	[PM_USB_STATE_ON] = PM_CAP_WAKEUP | PM_CAP_ACCESS | PM_CAP_CONTEXT |
@@ -158,7 +140,7 @@ static PmWakeEventGicProxy pmUsb0Wake = {
 	.group = 2U,
 };
 
-static u32 PmUsbPowers[] = {
+static u8 PmUsbPowers[] = {
 	[PM_USB_STATE_UNUSED] = DEFAULT_USB_POWER_OFF,
 	[PM_USB_STATE_OFF] = DEFAULT_USB_POWER_OFF,
 	[PM_USB_STATE_ON] = DEFAULT_USB_POWER_ON,
