+++ lib/sw_apps/zynqmp_pmufw/src/xpfw_error_manager.c
@@ -1,26 +1,8 @@
 /******************************************************************************
-* Copyright (C) 2015 - 2019 Xilinx, Inc.  All rights reserved.
-*
-* Permission is hereby granted, free of charge, to any person obtaining a copy
-* of this software and associated documentation files (the "Software"), to deal
-* in the Software without restriction, including without limitation the rights
-* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-* copies of the Software, and to permit persons to whom the Software is
-* furnished to do so, subject to the following conditions:
-*
-* The above copyright notice and this permission notice shall be included in
-* all copies or substantial portions of the Software.
-*
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
-* THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-* THE SOFTWARE.
-*
-*
+* Copyright (c) 2015 - 2020 Xilinx, Inc.  All rights reserved.
+* SPDX-License-Identifier: MIT
 ******************************************************************************/
+
 #include "xpfw_error_manager.h"
 #include "xpfw_mod_em.h"
 #include "xpfw_xpu.h"
@@ -39,16 +21,16 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NULL,
 			.Type = (u8)0U,
 			.Action = EM_ACTION_NONE,
-			.ChngPerm = (u32)EM_ACTION_CHANGE_PERM_NONE
+			.ChngPerm = (u16)EM_ACTION_CHANGE_PERM_NONE
 	},
 	[EM_ERR_ID_CSU_ROM] = {
 			.RegMask = PMU_GLOBAL_ERROR_STATUS_2_CSU_ROM_MASK,
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_2,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_PMU_PB] =
 	{
@@ -56,9 +38,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_2,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_PMU_SERVICE] =
 	{
@@ -66,9 +48,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_2,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_PMU_FW] =
 	{
@@ -76,9 +58,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_2,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_PMU_UC] =
 	{
@@ -86,9 +68,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_2,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_CSU] =
 	{
@@ -96,9 +78,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_2,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_PLL_LOCK] =
 	{
@@ -106,9 +88,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_2,
 			.Action = EM_ACTION_NONE,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_PL] =
 	{
@@ -116,9 +98,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_2,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_TO] =
 	{
@@ -126,9 +108,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_2,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_AUX3] =
 	{
@@ -136,9 +118,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_AUX2] =
 	{
@@ -146,9 +128,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_AUX1] =
 	{
@@ -156,9 +138,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_AUX0] =
 	{
@@ -166,9 +148,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_CSU_SWDT] =
 	{
@@ -176,7 +158,7 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_SRST,
-			.ChngPerm = (u32)EM_ACTION_CHANGE_PERM_NONE
+			.ChngPerm = (u16)EM_ACTION_CHANGE_PERM_NONE
 	},
 	[EM_ERR_ID_CLK_MON] =
 	{
@@ -184,9 +166,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_XMPU] =
 	{
@@ -194,8 +176,8 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = XPfw_XpuIntrHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_CUSTOM,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_PWR_SUPPLY] =
 	{
@@ -203,9 +185,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_FPD_SWDT] =
 	{
@@ -213,9 +195,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = SwdtHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_SRST,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_LPD_SWDT] =
 	{
@@ -223,8 +205,8 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = SwdtHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_SRST,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_RPU_CCF] =
 	{
@@ -232,8 +214,8 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_RPU_LS] =
 	{
@@ -241,8 +223,8 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = RpuLsHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_CUSTOM,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_FPD_TEMP] =
 	{
@@ -250,9 +232,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_LPD_TEMP] =
 	{
@@ -260,8 +242,8 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_RPU1] =
 	{
@@ -269,8 +251,8 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_RPU0] =
 	{
@@ -278,8 +260,8 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_OCM_ECC] =
 	{
@@ -287,9 +269,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 	[EM_ERR_ID_DDR_ECC] =
 	{
@@ -297,9 +279,9 @@ struct XPfw_Error_t ErrorTable[EM_ERR_ID_MAX] = {
 			.Handler = NullHandler,
 			.Type = EM_ERR_TYPE_1,
 			.Action = EM_ACTION_PSERR,
-			.ChngPerm = (u32)(IPI_PMU_0_IER_APU_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_0_MASK) |
-						(u32)(IPI_PMU_0_IER_RPU_1_MASK)
+			.ChngPerm = (u16)(IPI_PMU_0_IER_APU_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_0_MASK) |
+						(u16)(IPI_PMU_0_IER_RPU_1_MASK)
 	},
 };
